pyproject.toml
[tool.poetry]
name = "myvinted"
version = "0.1.0"
description = ""
authors = ["Sander van den Oord <SandervandenOord@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pydantic = "^2.5.3"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
